Cast double to int for GoogleSheetInterval.

+++ src/UIState/SetGoogleSheetInterval.cpp
@@ -8,10 +8,10 @@
 #include "../Devices/LiquidCrystal_TC.h"
 
 void SetGoogleSheetInterval::setValue(double value) {
-  EEPROM_TC::instance()->setGoogleSheetInterval(value);
+  EEPROM_TC::instance()->setGoogleSheetInterval((int)value);
 
   char output[17];
-  sprintf(output, "New interval=%i", value);
+  sprintf(output, "New interval=%i", (int)value);
   LiquidCrystal_TC::instance()->writeLine(output, 1);
   delay(1000);  // 1 second
   returnToMainMenu();
